+++ app/helpers/application_helper.rb
@@ -7,4 +7,16 @@ module ApplicationHelper
 
     link_to sha[0..7], COMMIT_STEM + sha
   end
+
+  def title(research_session, step)
+    if research_session.nil?
+      I18n.t('application.title')
+    elsif research_session.new_record?
+      "#{I18n.t('application.create_new_form')} – #{I18n.t('application.title')}"
+    elsif research_session.status == 'incentive' && step.nil?
+      "Preview – #{research_session.name.strip}"
+    else
+      "#{step.to_s.humanize} – #{research_session.name.strip}"
+    end
+  end
 end